refac: file validator를 유틸로 분리
