+++ src/main/kotlin/com/novel/cloud/web/domain/file/controller/FileController.kt
@@ -2,6 +2,7 @@ package com.novel.cloud.web.domain.file.controller
 
 import com.novel.cloud.web.domain.file.service.S3UploadService
 import com.novel.cloud.web.path.ApiPath
+import com.novel.cloud.web.utils.FileValidateUtils
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.web.bind.annotation.PostMapping
@@ -17,7 +18,10 @@ class FileController(
 
     @Operation(summary = "에디터 이미지 단일 업로드")
     @PostMapping(ApiPath.FILE_UPLOAD)
-    fun editorImageUpload(@RequestPart(value = "image") image: MultipartFile): String {
+    fun editorImageUpload(
+        @RequestPart(value = "image") image: MultipartFile
+    ): String {
+        FileValidateUtils.supportedFileValidationCheck(image)
         return s3UploadService.upload(image)
     }
