+++ src/main/kotlin/com/novel/cloud/web/exception/FileException.kt
@@ -1,0 +1,7 @@
+package com.novel.cloud.web.exception
+
+import org.springframework.http.HttpStatus
+
+class FileException (
+    message: String
+) : GeneralHttpException(HttpStatus.BAD_REQUEST, message, null)
