--- src/main/kotlin/com/novel/cloud/web/validator/VaildImage.kt
@@ -1,14 +1,0 @@
-package com.novel.cloud.web.validator
-
-import javax.validation.Constraint
-import javax.validation.Payload
-import kotlin.reflect.KClass
-
-@Target(AnnotationTarget.FIELD, AnnotationTarget.VALUE_PARAMETER)
-@Retention(AnnotationRetention.RUNTIME)
-@Constraint(validatedBy = [ImageFileValidator::class])
-annotation class ValidImage(
-    val message: String = "올바른 이미지 파일이 아닙니다.",
-    val groups: Array<KClass<*>> = [],
-    val payload: Array<KClass<out Payload>> = []
-)
